Extract interface for fs watching
Allows for adding additional "drivers" for
filesystem monitoring.
