+++ src/Filesystem/ChangesListener.php
@@ -1,35 +1,13 @@
-<?php declare(strict_types=1);
+<?php
 
+declare(strict_types=1);
 namespace seregazhuk\PhpWatcher\Filesystem;
 
-use Evenement\EventEmitter;
-use React\ChildProcess\Process;
-use React\EventLoop\LoopInterface;
-use RuntimeException;
 use seregazhuk\PhpWatcher\Config\WatchList;
 
-final class ChangesListener extends EventEmitter
+interface ChangesListener
 {
-    private const INTERVAL = 0.15;
+    public function start(WatchList $watchList): void;
 
-    private $loop;
-
-    public function __construct(LoopInterface $loop)
-    {
-        $this->loop = $loop;
-    }
-
-    public function start(WatchList $watchList): void
-    {
-        $watcher = ResourceWatcherBuilder::create($watchList);
-
-        $this->loop->addPeriodicTimer(
-            self::INTERVAL,
-            function () use ($watcher) {
-                if ($watcher->findChanges()->hasChanges()) {
-                    $this->emit('change');
-                }
-            }
-        );
-    }
+    public function onChange(callable $callback): void;
 }
