+++ tests/Feature/ChangesListenerTest.php
@@ -5,10 +5,11 @@ namespace seregazhuk\PhpWatcher\Tests\Feature;
 use PHPUnit\Framework\TestCase;
 use React\EventLoop\Factory;
 use seregazhuk\PhpWatcher\Config\WatchList;
-use seregazhuk\PhpWatcher\Filesystem\ChangesListener;
+use seregazhuk\PhpWatcher\Filesystem\ResourceWatcherBased\ChangesListener;
 use seregazhuk\PhpWatcher\Tests\Feature\Helper\Filesystem;
 use seregazhuk\PhpWatcher\Tests\Feature\Helper\WithFilesystem;
-use function \Clue\React\Block\sleep;
+
+use function Clue\React\Block\sleep;
 
 final class ChangesListenerTest extends TestCase
 {
